将欧拉编码器和球面编码器的输出数据类型更改为float32

+++ SPEN/pose/euler.py
@@ -22,7 +22,7 @@ class EulerEncoder():
         euler = rotation.as_euler("YXZ", degrees=False)
 
         return {
-            "euler": euler
+            "euler": euler.astype(np.float32)
         }
